+++ cpp/include/gigperformer/sdk/GigPerformerFunctions.h
@@ -176,6 +176,10 @@ class GigPerformerFunctions
     /// \brief   Query GP for the index of the currently selected song.
     int getCurrentSongIndex();
 
+    /// \brief   Query GP for the name of the variation associated with the song and song part with the given
+    ///          (zero-based) indices.
+    std::string getVariationNameForSongPart(int atSongIndex, int atPartIndex);
+
     /// \brief   Query GP for the number of song parts in the song at the given index.
     int getSongpartCount(int atSongIndex);